+++ src/mergeDeep.test.ts
@@ -1,8 +1,5 @@
+import { doNothing } from "./doNothing";
 import { mergeDeep } from "./mergeDeep";
-
-const NOOP = (): void => {
-  /* do nothing */
-};
 
 describe("runtime (dataFirst)", () => {
   test("should merge objects", () => {
@@ -49,10 +46,10 @@ describe("runtime (dataFirst)", () => {
   });
 
   it("should work with weird object types, functions", () => {
-    const a = { foo: NOOP };
+    const a = { foo: doNothing() };
     const b = { foo: 123 };
     expect(mergeDeep(a, b)).toEqual({ foo: 123 });
-    expect(mergeDeep(b, a)).toEqual({ foo: NOOP });
+    expect(mergeDeep(b, a)).toEqual({ foo: doNothing() });
   });
 
   it("should work with weird object types, date", () => {
@@ -80,55 +77,5 @@ describe("runtime (dataLast)", () => {
     const a = { foo: "baz", x: 1 };
     const b = { foo: "bar", y: 2 };
     expect(mergeDeep(b)(a)).toEqual({ foo: "bar", x: 1, y: 2 });
-  });
-});
-
-describe("typing", () => {
-  it("trivially merges disjoint objects", () => {
-    const a = { foo: "bar" };
-    const b = { bar: "baz" };
-    const result = mergeDeep(a, b);
-    expectTypeOf(result).toEqualTypeOf<{ foo: string; bar: string }>();
-  });
-
-  it("merges fully overlapping types", () => {
-    const a = { foo: "bar" };
-    const b = { foo: "baz" };
-    const result = mergeDeep(a, b);
-    expectTypeOf(result).toEqualTypeOf<{ foo: string }>();
-  });
-
-  it("merges semi-overlapping types", () => {
-    const a = { foo: "bar", x: 1 };
-    const b = { foo: "baz", y: 2 };
-    const result = mergeDeep(a, b);
-    expectTypeOf(result).toEqualTypeOf<{ foo: string; x: number; y: number }>();
-  });
-
-  it("deeply merges", () => {
-    const a = { foo: { bar: "baz" } };
-    const b = { foo: { qux: "quux" } };
-    const result = mergeDeep(a, b);
-    expectTypeOf(result).toEqualTypeOf<{ foo: { bar: string; qux: string } }>();
-  });
-
-  it("overrides types", () => {
-    const a = { foo: { bar: "baz" } };
-    const b = { foo: "qux" };
-    expectTypeOf(mergeDeep(a, b)).toEqualTypeOf<typeof b>();
-    expectTypeOf(mergeDeep(b, a)).toEqualTypeOf<typeof a>();
-  });
-
-  it("doesn't spread arrays", () => {
-    const a = { foo: ["bar"] as const };
-    const b = { foo: ["baz"] as const };
-    const result = mergeDeep(a, b);
-    expectTypeOf(result).toEqualTypeOf<{ foo: readonly ["baz"] }>();
-  });
-
-  it("doesn't recurse into arrays", () => {
-    const a = { foo: [{ bar: "baz" }] };
-    const b = { foo: [{ bar: "hello, world" }] };
-    expectTypeOf(mergeDeep(a, b)).toEqualTypeOf<typeof b>();
   });
 });
